Fix some C++11 convertions errors

+++ InputSettings.cpp
@@ -509,7 +509,8 @@ CInputSettings::InputInfo CInputSettings::ms_ainputinfo[CInputSettings::NumInput
       "Fire",
       RSP_SK_CONTROL,
       0,
-      RSP_JOY_BUT_17,
+      //FixMe: I don't think this works, as the value will be zero after the cast
+      static_cast<short>(RSP_JOY_BUT_17),
     },
     {
       "Fire",
@@ -551,7 +552,7 @@ CInputSettings::InputInfo CInputSettings::ms_ainputinfo[CInputSettings::NumInput
       "Duck",
       RSP_SK_F,
       0,
-      RSP_JOY_BUT_16,
+      static_cast<short>(RSP_JOY_BUT_16),
     },
     {
       "Rejuvenate",
